💥 breaking change: move away from `SetRouteSpecificData`, incorporate it into `RouteSegment`
This change moves towards `Route` as a list of `RouteSegement`s.
Each segment has an `id` (previously `RouteElementIdentifier`) and `args` (previously set via
`SetRouteSpecificData`). That means that the route specific data is set directly via `SetRouteAction`
as well as dispatched to `Routable`s in all `push`, `pop` and `change` calls.